+++ src/PhpWord/Settings.php
@@ -14,12 +14,25 @@ namespace PhpOffice\PhpWord;
  */
 class Settings
 {
+    /**    Available Zip library classes */
+    const PCLZIP     = 'PhpOffice\\PhpWord\\Shared\\ZipArchive';
+    const ZIPARCHIVE = 'ZipArchive';
+
     /**
      * Compatibility option for XMLWriter
      *
      * @var boolean
      */
     private static $_xmlWriterCompatibility = true;
+
+    /**
+     * Name of the class used for Zip file management
+     *  e.g.
+     *      ZipArchive
+     *
+     * @var string
+     */
+    private static $_zipClass = self::ZIPARCHIVE;
 
     /**
      * Set the compatibility option used by the XMLWriter
@@ -45,4 +58,34 @@ class Settings
     {
         return self::$_xmlWriterCompatibility;
     }
+
+    /**
+     * Set the Zip handler Class that PHPWord should use for Zip file management (PCLZip or ZipArchive)
+     *
+     * @param  string $zipClass  The Zip handler class that PHPWord should use for Zip file management
+     *   e.g. Settings::PCLZip or Settings::ZipArchive
+     * @return boolean Success or failure
+     */
+    public static function setZipClass($zipClass)
+    {
+        if (($zipClass === self::PCLZIP) ||
+            ($zipClass === self::ZIPARCHIVE)) {
+            self::$_zipClass = $zipClass;
+            return true;
+        }
+        return false;
+    } // function setZipClass()
+
+    /**
+     * Return the name of the Zip handler Class that PHPWord is configured to use (PCLZip or ZipArchive)
+     *  or Zip file management
+     *
+     * @return string Name of the Zip handler Class that PHPWord is configured to use
+     *  for Zip file management
+     *  e.g. Settings::PCLZip or Settings::ZipArchive
+     */
+    public static function getZipClass()
+    {
+        return self::$_zipClass;
+    } // function getZipClass()
 }