+++ src/PhpWord/Writer/ODText.php
@@ -12,6 +12,7 @@ namespace PhpOffice\PhpWord\Writer;
 use PhpOffice\PhpWord\Exceptions\Exception;
 use PhpOffice\PhpWord\PhpWord;
 use PhpOffice\PhpWord\HashTable;
+use PhpOffice\PhpWord\Settings;
 use PhpOffice\PhpWord\Writer\ODText\Content;
 use PhpOffice\PhpWord\Writer\ODText\Manifest;
 use PhpOffice\PhpWord\Writer\ODText\Meta;
@@ -26,13 +27,13 @@ class ODText extends Writer implements IWriter
     /**
      * Private unique PHPWord_Worksheet_BaseDrawing HashTable
      *
-     * @var \PhpOffice\PhpWord\HashTable
+     * @var HashTable
      */
     private $drawingHashTable;
 
     /**
      * Create new ODText writer
-     * @param \PhpOffice\PhpWord\PhpWord $phpWord
+     * @param PhpWord $phpWord
      */
     public function __construct(PhpWord $phpWord = null)
     {
@@ -57,7 +58,7 @@ class ODText extends Writer implements IWriter
      * Save PhpWord to file
      *
      * @param  string $pFilename
-     * @throws \PhpOffice\PhpWord\Exceptions\Exception
+     * @throws Exception
      */
     public function save($pFilename = null)
     {
@@ -65,11 +66,23 @@ class ODText extends Writer implements IWriter
             $pFilename = $this->getTempFile($pFilename);
 
             // Create new ZIP file and open it for writing
-            $objZip = new \ZipArchive();
+            $zipClass = Settings::getZipClass();
+            $objZip = new $zipClass();
+
+            // Retrieve OVERWRITE and CREATE constants from the instantiated zip class
+            // This method of accessing constant values from a dynamic class should work with all appropriate versions of PHP
+            $ro = new \ReflectionObject($objZip);
+            $zipOverWrite = $ro->getConstant('OVERWRITE');
+            $zipCreate = $ro->getConstant('CREATE');
+
+            // Remove any existing file
+            if (file_exists($pFilename)) {
+                unlink($pFilename);
+            }
 
             // Try opening the ZIP file
-            if ($objZip->open($pFilename, \ZipArchive::OVERWRITE) !== true) {
-                if ($objZip->open($pFilename, \ZipArchive::CREATE) !== true) {
+            if ($objZip->open($pFilename, $zipOverWrite) !== true) {
+                if ($objZip->open($pFilename, $zipCreate) !== true) {
                     throw new Exception("Could not open " . $pFilename . " for writing.");
                 }
             }
@@ -101,7 +114,8 @@ class ODText extends Writer implements IWriter
                         $imagePath = substr($imagePath, 6);
                         $imagePathSplitted = explode('#', $imagePath);
 
-                        $imageZip = new \ZipArchive();
+                        $zipClass = Settings::getZipClass();
+                        $imageZip = new $zipClass();
                         $imageZip->open($imagePathSplitted[0]);
                         $imageContents = $imageZip->getFromName($imagePathSplitted[1]);
                         $imageZip->close();
@@ -139,7 +153,7 @@ class ODText extends Writer implements IWriter
     /**
      * Get PHPWord_Worksheet_BaseDrawing HashTable
      *
-     * @return \PhpOffice\PhpWord\HashTable
+     * @return HashTable
      */
     public function getDrawingHashTable()
     {